add max clients and tunnels

+++ src/server/configuration.rs
@@ -69,8 +69,8 @@ pub struct ServerConfiguration {
     pub monitor_key: Option<String>,
     pub endpoints: HashMap<String, EndpointConfiguration>,
     pub encryption: ServerEncryption,
-    pub max_tunnels: u16,
-    pub max_clients: u16,
+    pub max_tunnels: usize,
+    pub max_clients: usize,
     pub max_links: u16,
 }
